requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
argparse>=1.4.0
jinja2>=3.1.2
beautifulsoup4>=4.12.0
markdown>=3.4.0
retrying>=1.3.4
pytz>=2023.3
